requirements.txt
# Database
psycopg[binary]

# Document processing
PyMuPDF
python-docx
python-dotenv

# Web UI
streamlit
pandas

# Environment and utilities
python-dotenv

# Summarization pipeline
dspy-ai
sentence-transformers>=2.2.0
requests
numpy
scikit-learn

# Embeddings and transformers
pgvector
torch
transformers
sentencepiece
protobuf

# Text processing
langchain-text-splitters

# RAG DEMO
google.generativeai
psycopg_pool
torchvision

